Move data sources into separate module
