+++ qtgui/widgets/inputselector.py
@@ -1,255 +1,15 @@
-from os import listdir, stat
-from os.path import isfile, isdir, join, basename
-from random import randint
-
-import numpy as np
-from scipy.misc import imread
-
-from collections import namedtuple
-
-from PyQt5.QtWidgets import QWidget, QFileDialog
-
-# FIXME[todo]: add docstrings!
-
-InputData = namedtuple('Data', ['data', 'name'])
-
-class DataSource:
-    '''An abstract base class for different types of data sources.  The
-    individual elements of a data source can be accessed using an array like
-    notation.
-
-    Attributes
-    ----------
-    _description    :   str
-                        Short description of the dataset
-
-    '''
-    _description: str = None
-
-    def __init__(self, description=None):
-        '''Create a new DataSource
-
-        Parameters
-        ----------
-        description :   str
-                        Description of the dataset
-        '''
-        self._description = description
-
-    def __getitem__(self, index: int):
-        '''Provide access to the records in this data source.'''
-        pass
-
-    def __len__(self):
-        '''Get the number of entries in this data source.'''
-        pass
-
-    def getDescription(self) -> str:
-        '''Get the description for this DataSource'''
-        return self._description
-
-
-class DataArray(DataSource):
-    '''A ``DataArray`` the stores all entries in an array (like the MNIST
-    character data). That means that all entries will have the same sizes.
-
-    Attributes
-    ----------
-    _array  :   np.ndarray
-                An array of input data. Can be None.
-    '''
-    _array: np.ndarray = None
-
-    def __init__(self, array: np.ndarray=None, description: str=None):
-        '''Create a new DataArray
-
-        Parameters
-        ----------
-        array   :   np.ndarray
-                    Numpy data array
-        description :   str
-                        Description of the data set
-
-        '''
-        super().__init__(description)
-        if array is not None:
-            self.setArray(array, description)
-
-    def setArray(self, array, description='array'):
-        '''Set the array of this DataSource.
-
-        Parameters
-        ----------
-        array   :   np.ndarray
-                    Numpy data array
-        description :   str
-                        Description of the data set
-        '''
-        self._array = array
-        self._description = description
-
-    def __getitem__(self, index: int):
-        if self._array is None or index is None:
-            return None, None
-        data = self._array[index]
-        info = 'Image ' + str(index) + ' from ' + self._description
-        return InputData(data, info)
-
-    def __len__(self):
-        if self._array is None:
-            return 0
-        return len(self._array)
-
-
-class DataFile(DataArray):
-    '''Data source for reading from a file.
-
-    Attributes
-    ----------
-    _filename   :   str
-                    The name of the file from which the data are read.
-    '''
-
-    _filename: str = None
-
-    def __init__(self, filename: str=None):
-        '''Create a new data file.
-
-        Parameters
-        ----------
-        filename    :   str
-                        Name of the file containing the data
-        '''
-        super().__init__()
-        if filename is not None:
-            self.setFile(filename)
-
-    def setFile(self, filename: str):
-        '''Set the data file to be used.
-
-        Parameters
-        ----------
-        filename    :   str
-                        Name of the file containing the data
-        '''
-        self._filename = filename
-        data = np.load(filename, mmap_mode='r')
-        self.setArray(data, basename(self._filename))
-
-    def getFile(self) -> str:
-        '''Get the underlying file name'''
-        return self._filename
-
-    def selectFile(self, parent: QWidget=None):
-        filters = 'Numpy Array (*.npy);; All Files (*)'
-        filename, _ = QFileDialog.getOpenFileName(
-            parent,
-            'Select input data archive',
-            self._filename,
-            filters
-        )
-        if filename is None or not isfile(filename):
-            raise FileNotFoundError('The specified file %s could not be found.' %
-                                    filename)
-        self.setFile(filename)
-
-
-class DataSet(DataArray):
-
-    def __init__(self, name: str = None):
-        super().__init__()
-        self.load(name)
-
-    def load(self, name: str):
-        if name == 'mnist':
-            from keras.datasets import mnist
-            data = mnist.load_data()[0][0]
-            self.setArray(data, 'MNIST')
-        else:
-            raise ValueError('Unknown dataset: {}'.format(name))
-
-    def getName(self) -> str:
-        return 'MNIST'
-
-
-class DataDirectory(DataSource):
-    '''A data directory contains data entries (e.g., images), in
-    individual files. Each files is only read when accessed.
-
-    Attributes
-    ----------
-    _dirname    :   str
-                    A directory containing input data files. Can be None.
-
-
-    _filenames  :   list
-                    A list of filenames in the data directory. Will be None if
-                    no directory was selected. An empty list indicates that no
-                    suitable files where found in the directory.
-    '''
-    _dirname: str = None
-    _filenames: list = None
-
-    def __init__(self, dirname: str=None):
-        '''Create a new DataDirectory
-
-        Parameters
-        ----------
-        dirname :   str
-                    Name of the directory with files
-        '''
-        super().__init__()
-        self.setDirectory(dirname)
-
-    def setDirectory(self, dirname: str):
-        '''Set the directory to load from
-
-        Parameters
-        ----------
-        dirname :   str
-                    Name of the directory with files
-        '''
-        self._dirname = dirname
-        if self._dirname is None:
-            self._filenames = None
-        else:
-            self._filenames = [f for f in listdir(self._dirname)
-                               if isfile(join(self._dirname, f))]
-
-    def getDirectory(self) -> str:
-        return self._dirname
-
-    def __getitem__(self, index):
-        if not self._filenames:
-            return None, None
-        else:
-            # TODO: This can be much improved by caching and/or prefetching
-            filename = self._filenames[index]
-            data = imread(join(self._dirname, filename))
-            return InputData(data, filename)
-
-    def __len__(self):
-        if not self._filenames:
-            return 0
-        return len(self._filenames)
-
-    def selectDirectory(self, parent: QWidget=None):
-        dirname = QFileDialog.getExistingDirectory(parent, 'Select Directory')
-        if not dirname or not isdir(dirname):
-            raise FileNotFoundError('%s is not a directory.' % dirname)
-        else:
-            self.setDirectory(dirname)
-
-
 from PyQt5.QtCore import Qt, QSize, pyqtSignal
 from PyQt5.QtGui import QFontMetrics, QIntValidator, QIcon
 from PyQt5.QtWidgets import QWidget, QPushButton, QRadioButton, QLineEdit
 from PyQt5.QtWidgets import QLabel, QGroupBox
 from PyQt5.QtWidgets import QHBoxLayout, QVBoxLayout, QSizePolicy
 
-from observer import Observer
+import observer
+import numpy as np
 
-class QInputSelector(QWidget, Observer):
+from qtgui.datasources import DataArray, DataFile, DataDirectory
+
+class QInputSelector(QWidget, observer.Observer):
     '''A Widget to select input data (probably images).  There are
     different modes of selection: from an array, from a file, from a
     directory or from some predefined dataset.
